refactor: reorganize component file structure
Move CustomCard and ImagePlaceholder components to the UI folder.
