+++ components/ui/CustomCard.tsx
@@ -1,8 +1,8 @@
 "use client";
 
 import React from 'react';
-import { Card, CardHeader, CardContent, CardFooter, CardTitle, CardDescription } from '../ui/card';
-import { Button } from '../ui/button';
+import { Card, CardHeader, CardContent, CardFooter, CardTitle, CardDescription } from './card';
+import { Button } from './button';
 
 type CustomCardProps = {
   title: string;
