+++ components/ui/ImagePlaceholder.tsx
